+++ datanode/gateway/graphql/marshallers/marshallers.go
@@ -406,3 +406,23 @@ func MarshalValidatorStatus(s vega.ValidatorNodeStatus) graphql.Marshaler {
 func UnmarshalValidatorStatus(v interface{}) (vega.ValidatorNodeStatus, error) {
 	return vega.ValidatorNodeStatus_VALIDATOR_NODE_STATUS_UNSPECIFIED, ErrUnimplemented
 }
+
+func MarshalProtocolUpgradeProposalStatus(s eventspb.ProtocolUpgradeProposalStatus) graphql.Marshaler {
+	return graphql.WriterFunc(func(w io.Writer) {
+		w.Write([]byte(strconv.Quote(s.String())))
+	})
+}
+
+func UnmarshalProtocolUpgradeProposalStatus(v interface{}) (eventspb.ProtocolUpgradeProposalStatus, error) {
+	s, ok := v.(string)
+	if !ok {
+		return eventspb.ProtocolUpgradeProposalStatus_PROTOCOL_UPGRADE_PROPOSAL_STATUS_UNSPECIFIED, fmt.Errorf("expected proposal type in force to be a string")
+	}
+
+	t, ok := eventspb.ProtocolUpgradeProposalStatus_value[s] // v2.ListGovernanceDataRequest_Type_value[s]
+	if !ok {
+		return eventspb.ProtocolUpgradeProposalStatus_PROTOCOL_UPGRADE_PROPOSAL_STATUS_UNSPECIFIED, fmt.Errorf("failed to convert proposal type from GraphQL to Proto: %v", s)
+	}
+
+	return eventspb.ProtocolUpgradeProposalStatus(t), nil
+}